refactor: 자바스크립트 파일 모듈화 및 상수화
