+++ index.html
@@ -48,6 +48,6 @@
       </button>
       <main class="grid"></main>
     </div>
-    <script type="module" src="main.js"></script>
+    <script type="module" src="./module/app.js"></script>
   </body>
 </html>
